requirements.txt
celery
redis
Django
djangorestframework
drf-spectacular
django-cors-headers
djangorestframework-simplejwt
